feat: enable cross-platform zipapp packaging for CLI
Extract _build_index from mcp_app.py into index_builder.py to decouple
the CLI from heavy MCP dependencies (fastmcp, pydantic, cryptography).
The CLI now only needs pure-Python deps (click, pyyaml, pathspec),
enabling packaging as a cross-platform .pyz zipapp via build_zipapp.sh.

Bump version to 0.4.37.
